Refactor the SMC kernels
